LA-109 Move CLI param handling to separate class. (There are more params to be added later.)
